[AAP-46162] create lightspeed-inline-agent
issue: https://issues.redhat.com/browse/AAP-46162
This will allow to filter the used tools before submitted to llm usage.
The principe used here is to submit the tools to llm to filter the tools before it gets analysed by the main task that can be stuck when the number of tools is very big, in our usage the number is more than 800 tools only to count AAP controller with gateway apis.

Signed-off-by: Djebran Lezzoum <ldjebran@gmail.com>